Extract Tor configuration to separate class
